OcdFileExport: Fix tabulator type conversion
The export now converts the tabulator position to a 32 bit signed
value since the convertSize function that was used before converted it
to a 16 bit value instead (which was indeed only relevant for tab
values larger than 327,67mm).

+++ src/fileformats/ocd_file_export.cpp
@@ -1961,7 +1961,7 @@ void OcdFileExport::setupTextSymbolSpecial(const TextSymbol* text_symbol, OcdTex
 		           .arg(text_symbol->getPlainTextName()).arg(max_tabs));
 	}
 	for (auto i = 0u; i < ocd_text_special.num_tabs; ++i)
-		ocd_text_special.tab_pos[i] = convertSize(text_symbol->getCustomTab(i));
+		ocd_text_special.tab_pos[i] = convertSize(qint64(text_symbol->getCustomTab(i)));
 }
